(CU-85zt1pbdz ui) cleanup: helpers classes and funcs -> helpers.py; type hints & style
